services/app/requirements.txt
email-validator==1.3.0
flask==2.2.2 
flask-cors==3.0.10 
flask-marshmallow==0.14.0 
flask-migrate==4.0.0 
flask-sqlalchemy==3.0.2 
flask-wtf==1.0.1
marshmallow-sqlalchemy==0.28.1
psycopg2-binary==2.9.5 
python-dotenv==0.21.0 
sqlalchemy-utils==0.38.3
flask-bcrypt==1.0.1 
flask-login==0.6.2
redis==4.4.0
celery==5.2.7
boto3==1.26.33
gunicorn==20.1.0
python-json-logger==2.0.4
python-logstash==0.4.8
blinker==1.5
flask-dance==6.2.0
pillow==9.4.0
flask-mail==0.9.1